+++ src/Config/Repository.php
@@ -68,11 +68,12 @@ final class Repository extends AbstractRepository
     protected function getFields() : array
     {
         return [
-            FormBuilder::KEY_LINK_OBJECTS                             => [Config::TYPE_BOOLEAN, true],
+            FormBuilder::KEY_LINK_CONTAINER_OBJECTS                   => [Config::TYPE_BOOLEAN, true],
             FormBuilder::KEY_MAX_DEEP_METHOD                          => [Config::TYPE_INTEGER, TreeRepository::MAX_DEEP_METHOD_END],
             FormBuilder::KEY_MAX_DEEP_METHOD_START_HIDE               => [Config::TYPE_BOOLEAN, false],
             FormBuilder::KEY_OBJECT_TYPES                             => [Config::TYPE_JSON, self::srContainerObjectTree()->tree()->getObjectTypes(), true],
             FormBuilder::KEY_ONLY_SHOW_CONTAINER_OBJECTS_IF_NOT_EMPTY => [Config::TYPE_BOOLEAN, false],
+            FormBuilder::KEY_OPEN_LINKS_IN_NEW_TAB                    => [Config::TYPE_BOOLEAN, true],
             FormBuilder::KEY_RECURSIVE_COUNT                          => [Config::TYPE_BOOLEAN, false]
         ];
     }